refactor: consolidate contents into src/data and update imports
